+++ src/Data/Morpheus/Execution/Server/Interpreter.hs
@@ -8,41 +8,24 @@ module Data.Morpheus.Execution.Server.Interpreter
   )
 where
 
-import           Data.Aeson                     ( encode )
-import           Data.ByteString                ( ByteString )
-import qualified Data.ByteString.Lazy.Char8    as LB
-                                                ( ByteString
-                                                , fromStrict
-                                                , toStrict
-                                                )
-import           Data.Text                      ( Text )
-import qualified Data.Text.Lazy                as LT
-                                                ( Text
-                                                , fromStrict
-                                                , toStrict
-                                                )
-import           Data.Text.Lazy.Encoding        ( decodeUtf8
-                                                , encodeUtf8
-                                                )
-import           Control.Monad.IO.Class         ( MonadIO() )
 
 -- MORPHEUS
+import           Data.Morpheus.Types.Internal.Subscription
+                                                ( Stream
+                                                , Input
+                                                , toOutStream
+                                                )
 import           Data.Morpheus.Execution.Server.Resolve
                                                 ( RootResCon
-                                                , byteStringIO
-                                                , statefulResolver
-                                                , statelessResolver
-                                                , streamResolver
                                                 , coreResolver
+                                                , statelessResolver
                                                 )
-import           Data.Morpheus.Execution.Server.Subscription
-                                                ( GQLState )
 import           Data.Morpheus.Types.Internal.Resolving
                                                 ( GQLRootResolver(..)
-                                                , ResponseStream
                                                 )
 import           Data.Morpheus.Types.IO         ( GQLRequest
                                                 , GQLResponse
+                                                , MapAPI(..)
                                                 )
 
 -- | main query processor and resolver
@@ -62,76 +45,20 @@ import           Data.Morpheus.Types.IO         ( GQLRequest
 --       -- or
 --       k :: GQLRequest -> IO GQLResponse
 --     @
-class Interpreter k m e  where
+class Interpreter e m a b where
   interpreter ::
        Monad m
-    => (RootResCon m e query mut sub) =>
-         GQLRootResolver m e query mut sub -> k
+    => (RootResCon m e query mut sub) 
+        => GQLRootResolver m e query mut sub
+        -> a
+        -> b
 
-{-
-  simple HTTP stateless Interpreter without side effects
--}
-type StateLess m a = a -> m a
-
-instance Interpreter (GQLRequest -> m GQLResponse) m e   where
+instance Interpreter e m GQLRequest (m GQLResponse) where
   interpreter = statelessResolver
 
-instance Interpreter (StateLess m LB.ByteString) m e  where
-  interpreter root = byteStringIO (statelessResolver root)
+instance Interpreter e m (Input api)  (Stream api e m) where
+  interpreter root = toOutStream (coreResolver root)
 
-instance Interpreter (StateLess m LT.Text) m e  where
-  interpreter root request =
-    decodeUtf8 <$> interpreter root (encodeUtf8 request)
-
-instance Interpreter (StateLess m ByteString) m e  where
-  interpreter root request =
-    LB.toStrict <$> interpreter root (LB.fromStrict request)
-
-instance Interpreter (StateLess m Text) m e  where
-  interpreter root request =
-    LT.toStrict <$> interpreter root (LT.fromStrict request)
-
-{-
-   HTTP Interpreter with state and side effects, every mutation will
-   trigger subscriptions in  shared `GQLState`
--}
-type WSPub m e a = GQLState m e -> a -> m a
-
-instance MonadIO m => Interpreter (WSPub m e LB.ByteString) m e where
-  interpreter root state = statefulResolver state (coreResolver root)
-
-instance MonadIO m => Interpreter (WSPub m e  LT.Text) m e where
-  interpreter root state request =
-    decodeUtf8 <$> interpreter root state (encodeUtf8 request)
-
-instance MonadIO m => Interpreter (WSPub m e  ByteString) m e where
-  interpreter root state request =
-    LB.toStrict <$> interpreter root state (LB.fromStrict request)
-
-instance MonadIO m => Interpreter (WSPub m e  Text) m e  where
-  interpreter root state request =
-    LT.toStrict <$> interpreter root state (LT.fromStrict request)
-
-{-
-   WebSocket Interpreter without state and side effects, mutations and subscription will return Actions
-   that will be executed in WebSocket server
--}
-type WSSub m e a = a -> ResponseStream e m a
-
-instance Interpreter (GQLRequest -> ResponseStream e m LB.ByteString) m e where
-  interpreter root request = encode <$> streamResolver root request
-
-instance Interpreter (WSSub m e  LB.ByteString) m e where
-  interpreter root = byteStringIO (streamResolver root)
-
-instance Interpreter (WSSub m e  LT.Text)  m e where
-  interpreter root request =
-    decodeUtf8 <$> interpreter root (encodeUtf8 request)
-
-instance Interpreter (WSSub m e ByteString) m e where
-  interpreter root request =
-    LB.toStrict <$> interpreter root (LB.fromStrict request)
-
-instance Interpreter (WSSub m e Text)  m e where
-  interpreter root request =
-    LT.toStrict <$> interpreter root (LT.fromStrict request)
+instance ( MapAPI a ) => 
+  Interpreter e m a (m a)  where
+    interpreter root = mapAPI (interpreter root)